+++ src/utilities/mergeCourses.ts
